+ IEEE-754 decimal128 (BID, 34 digits) reusing the shared decimal engine * DecimalFormat now derives parameters from the IEEE width formula (ForWidth); D32/D64/D128 are one parameterization * DecimalFloatMath encode/decode generalized to BigInteger raw (width-agnostic, fits 128-bit) * Decimal8/Decimal16 now do exact base-10 arithmetic via the engine (refit to their custom field) instead of binary double + Decimal128 spec tests + Decimal8/16 exact-arithmetic tests

+++ Corlib.Extensions/System/DataTypes/Decimal64.cs
@@ -36,7 +36,7 @@ public readonly struct Decimal64 : IComparable, IComparable<Decimal64>, IEquatab
 
   private Decimal64(ulong raw) => this.RawValue = raw;
 
-  private Decimal64(DecimalFloatMath.Value v) => this.RawValue = DecimalFloatMath.Encode(v, _fmt);
+  private Decimal64(DecimalFloatMath.Value v) => this.RawValue = (ulong)DecimalFloatMath.Encode(v, _fmt);
 
   private DecimalFloatMath.Value Value => DecimalFloatMath.Decode(this.RawValue, _fmt);
